Move future-related code in utils_pytorch.py into future
