+++ cleverhans/future/torch/tests/test_utils.py
@@ -2,7 +2,7 @@
 import numpy as np
 import torch
 
-from cleverhans import utils_pytorch
+import cleverhans.future.torch.utils as utils
 from cleverhans.devtools.checks import CleverHansTest
 
 class TestOptimizeLinear(CleverHansTest):
@@ -11,7 +11,7 @@ class TestOptimizeLinear(CleverHansTest):
   """
   def test_optimize_linear_linf(self):
     grad = torch.tensor([[1., -2.]])
-    eta = utils_pytorch.optimize_linear(grad, eps=1., norm=np.inf)
+    eta = utils.optimize_linear(grad, eps=1., norm=np.inf)
     objective = torch.sum(grad * eta)
 
     self.assertEqual(grad.size(), eta.size())
@@ -20,7 +20,7 @@ class TestOptimizeLinear(CleverHansTest):
 
   def test_optimize_linear_l2(self):
     grad = torch.tensor([[.5 ** .5, -.5 ** .5]])
-    eta = utils_pytorch.optimize_linear(grad, eps=1., norm=2)
+    eta = utils.optimize_linear(grad, eps=1., norm=2)
     objective = torch.sum(grad * eta)
 
     self.assertEqual(grad.size(), eta.size())
@@ -29,7 +29,7 @@ class TestOptimizeLinear(CleverHansTest):
 
   def test_optimize_linear_l1(self):
     grad = torch.tensor([[1., -2.]])
-    eta = utils_pytorch.optimize_linear(grad, eps=1., norm=1)
+    eta = utils.optimize_linear(grad, eps=1., norm=1)
     objective = torch.sum(grad * eta)
 
     self.assertEqual(grad.size(), eta.size())
@@ -38,7 +38,7 @@ class TestOptimizeLinear(CleverHansTest):
 
   def test_optimize_linear_l1_ties(self):
     grad = torch.tensor([[2., -2.]])
-    eta = utils_pytorch.optimize_linear(grad, eps=1., norm=1)
+    eta = utils.optimize_linear(grad, eps=1., norm=1)
     objective = torch.sum(grad * eta)
 
     self.assertEqual(grad.size(), eta.size())
